pyproject.toml
[project]
name = "transit-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka>=2.12.1",
    "gtfs-realtime-bindings>=1.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
